@@ -34,9 +34,9 @@ func init() {
 	config.RegisterResource(container.TypeSidecar, &container.Sidecar{}, &container.Provider{})
 	config.RegisterResource(copy.TypeCopy, &copy.Copy{}, &copy.Provider{})
 	config.RegisterResource(docs.TypeDocs, &docs.Docs{}, &docs.DocsProvider{})
-	config.RegisterResource(docs.TypeChapter, &docs.Chapter{}, &docs.ChapterProvider{})
+	config.RegisterResource(docs.TypeChapter, &docs.Chapter{}, &null.Provider{})
 	config.RegisterResource(docs.TypeTask, &docs.Task{}, &null.Provider{})
-	config.RegisterResource(docs.TypeBook, &docs.Book{}, &docs.BookProvider{})
+	config.RegisterResource(docs.TypeBook, &docs.Book{}, &null.Provider{})
 	config.RegisterResource(exec.TypeExec, &exec.Exec{}, &exec.Provider{})
 	config.RegisterResource(helm.TypeHelm, &helm.Helm{}, &helm.Provider{})
 	config.RegisterResource(ingress.TypeIngress, &ingress.Ingress{}, &ingress.Provider{})
